dom/svg/test/test_stroke-linecap-hit-testing.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=589648
-->
<head>
  <title>Test hit-testing of line caps</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="run()">
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function run() {
  var div = document.getElementById("div");
  var x = div.offsetLeft;
  var y = div.offsetTop;
  var got, expected;

  got = document.elementFromPoint(5 + x, 5 + y);
  expected = document.getElementById("zero-length-square-caps");
  is(got, expected, "Check hit on zero length subpath's square caps");

  SimpleTest.finish();
}

]]>
</script>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=500174">Mozilla Bug 500174</a>
<p id="display"></p>
<div id="content">

  <div width="100%" height="1" id="div"></div>
  <svg xmlns="http://www.w3.org/2000/svg" id="svg" width="400" height="300">
    <path id="zero-length-square-caps" stroke="blue" stroke-width="50"
          stroke-linecap="square" d="M25,25 L25,25"/>
  </svg>

</div>
<pre id="test">
</pre>
</body>
</html>
